+++ ndsampler/toypatterns.py
@@ -1,4 +1,0 @@
-"""
-Moved to kwcoco
-"""
-from kwcoco.toypatterns import *  # NOQA