+++ ndsampler/utils/util_futures.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+"""
+Note: this module also exists in kwcoco.utils
+"""
 from __future__ import absolute_import, division, print_function, unicode_literals
 import concurrent.futures
 from concurrent.futures import as_completed